+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * fix http bridge default tag prefix 'ws'
 * fix bridge connect error not close source connection
 * add command flags -print
+* fix http/websocket dialer not set default port
 
 # v0.0.1